Added first 'complete' draft of Katakana Practice
App will only run by going to terminal, going to location of app, and using 'python3 katakanaPractice.py' to get it to run. 

Features:
Will create a list using the keys provided from dictionary, and then will randomize, and will allow you to type the romanji then pass that into the dictionary to see if it matches the value of that key, and will tell you if it is right or wrong. 

v 0.01 (I'm not sure how to do versions yet)
This thing works, and will error check capitalization